refactor(plugins): Convert Events class to Typescript
